+++ static/js/stream_topic_history.js
@@ -1,5 +1,4 @@
 import {all_messages_data} from "./all_messages_data";
-import * as channel from "./channel";
 import {FoldDict} from "./fold_dict";
 import * as message_util from "./message_util";
 import * as stream_data from "./stream_data";
@@ -291,23 +290,8 @@ export function add_history(stream_id, server_history) {
     fetched_stream_ids.add(stream_id);
 }
 
-export function get_server_history(stream_id, on_success) {
-    if (fetched_stream_ids.has(stream_id)) {
-        on_success();
-        return;
-    }
-
-    const url = "/json/users/me/" + stream_id + "/topics";
-
-    channel.get({
-        url,
-        data: {},
-        success(data) {
-            const server_history = data.topics;
-            add_history(stream_id, server_history);
-            on_success();
-        },
-    });
+export function has_history_for(stream_id) {
+    return fetched_stream_ids.has(stream_id);
 }
 
 export function get_recent_topic_names(stream_id) {
